# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Readify API Documentation",
        "description": "API documentation for Readify application",
        "contact": {
            "email": "muhammadzelvan263@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Readify API Server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Retrieve list of books with optional filters",
                "operationId": "f31a4a8c234eea8c0b2fdcd6634c37ce",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Keyword for title, ISBN, or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Filter by author",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "publisher_id",
                        "in": "query",
                        "description": "Filter by publisher",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Only show available books",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Add a new book",
                "operationId": "3f946db73e3a02e8e2f56f60c398986b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "isbn",
                                    "publication_year",
                                    "total_copies",
                                    "available_copies",
                                    "author_ids",
                                    "category_ids"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "publication_year": {
                                        "type": "integer"
                                    },
                                    "total_copies": {
                                        "type": "integer"
                                    },
                                    "available_copies": {
                                        "type": "integer"
                                    },
                                    "cover_image": {
                                        "type": "string"
                                    },
                                    "publisher_id": {
                                        "type": "integer"
                                    },
                                    "book_shelf_id": {
                                        "type": "integer"
                                    },
                                    "author_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Fetch a specific book by ID",
                "operationId": "3184b7ae54e17e29fd5fa3027972790e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book found"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update existing book",
                "operationId": "63f8f46953b2690763bfbaa581da0668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book updated"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Remove a book by ID",
                "operationId": "c286320eba3b96b72b7baf6cd634ac1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
    },
...
}